@@ -6,8 +6,6 @@ use Symfony\Component\Validator\Constraints\Length;
 use App\Forms\MyConstraints\CzDate;
 use Symfony\Component\Validator\Constraints\IsTrue;
 use Symfony\Component\Validator\Mapping\ClassMetadata;
-use App\Entity\Models\AstromenModel;
-use App\Entity\Models\Db_wrap;
 use App\Utils\DateTools;
 
 class AstromanAdd {
@@ -16,14 +14,11 @@ class AstromanAdd {
     protected $dob;
     protected $skill;
     
-    /**
-     *
-     * @var Db_wrap
-     */
-    protected $db;
-    
-    public function __construct(Db_wrap $db) {
-        $this->db = $db;
+    public function __construct(
+            protected \App\Models\AstromenModel $astromen_model
+    )
+    {
+        
     }
 
     public function getFName() {
@@ -63,7 +58,7 @@ class AstromanAdd {
         $metadata->addPropertyConstraint('lName', new NotBlank(['message' => 'Zadejte příjmení']));
         $metadata->addPropertyConstraint('lName', new Length(['max' => 20, 'maxMessage' => 'Příjmení může mít nanejvýš 20 znaků']));
         $metadata->addPropertyConstraint('dob', new NotBlank(['message' => 'Zadejte datum narození']));
-        $metadata->addPropertyConstraint('dob', new CzDate(['message' => 'Neplatné datum narození']));
+//        $metadata->addPropertyConstraint('dob', new CzDate(['message' => 'Neplatné datum narození']));
         $metadata->addPropertyConstraint('skill', new NotBlank(['message' => 'Zadejte dovednost']));
         $metadata->addPropertyConstraint('skill', new Length(['max' => 45, 'maxMessage' => 'Dovednost může mít nanejvýš 45 znaků']));
         $metadata->addGetterConstraint('nameNotUsedYet', new IsTrue(['message' => 'Tento astronaut již existuje']));
@@ -71,10 +66,11 @@ class AstromanAdd {
     
     public function isNameNotUsedYet() {
         $dateTools = new DateTools;
-        if (empty($this->fName) or empty($this->lName) or empty($this->dob) or !$dateTools->checkCzDate($this->dob)) {
+        if (empty($this->fName) or empty($this->lName) or empty($this->dob)) {
             return true;
         }
-        $model = new AstromenModel($this->db);
+        
+        $model = $this->astromen_model;
         return !$model->isNameExists($this->fName, $this->lName, $this->dob);        
     }
 }
