@@ -1,8 +1,6 @@
 <?php
-
 namespace App\Forms\DataObjects;
-use App\Entity\Models\Db_wrap;
-use App\Entity\Models\AstromenModel;
+
 use Symfony\Component\Validator\Constraints\IsTrue;
 use Symfony\Component\Validator\Constraints\NotBlank;
 use App\Forms\MyConstraints\IsInt;
@@ -11,11 +9,6 @@ use App\Utils\StringTools;
 
 trait AstromanIdTrait {
     protected $id;
-    protected $db;
-    
-    public function __construct(Db_wrap $db) {
-        $this->db = $db;
-    }
 
     public function getId() {
     return $this->id;
@@ -37,7 +30,8 @@ trait AstromanIdTrait {
         if (!$stringTools->isInt($this->id)) {
             return true; //není-li $id int, nutné to podchytit jiným constrainem
         }
-        $model = new AstromenModel($this->db);
+        
+        $model = $this->astromen_model;
         return $model->idExists($this->id);
     }
 }
